+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Immutable\Sequence::snap()`
+- `Innmind\Immutable\Set::snap()`
+
+### Changed
+
+- `Innmind\Immutable\Set` implementation now directly uses `Innmind\Immutable\Sequence`
+
+### Deprecated
+
+- `Innmind\Immutable\Set::defer()`
+
 ## 5.13.0 - 2025-03-23
 
 ### Added